requirements.txt
streamlit
sentence-transformers
numpy
pandas
faiss-cpu
rank-bm25
pypdf
scikit-learn
